requirement.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
fastapi-redis-cache==0.2.5
Flask==2.0.2
greenlet==1.1.2
h11==0.12.0
idna==3.3
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
psycopg2-binary==2.9.1
pydantic==1.8.2
PyJWT==2.3.0
python-dateutil==2.8.2
PyYAML==6.0
redis==3.5.3
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
wincertstore==0.2
